Modificado protocolo de carga de Datos, sustituido por carga de elemento por elemento.

+++ Kilometros WebAPI/Models/HttpGet/DataController/DataDistanceResponse.cs
@@ -8,6 +8,6 @@ namespace Kilometros_WebAPI.Models.HttpGet.DataController {
         public DateTime Timestamp;
 
         public long Distance;
-        public int Steps;
+        public long Steps;
     }
 }
